requirements.txt
discord.py==2.5.2
twitchio==2.8.2
fastapi==0.115.6
uvicorn==0.34.0
aiohttp==3.9.5
python-dotenv==1.0.1
psutil==6.0.0
